+++ subsystems/posix/userland/windowd_appinfo.c
@@ -29,6 +29,7 @@ static const struct windowd_appinfo k_window_items[] = {
     {"Files", "/bin/filesapp", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(186, 128, 36)},
     {"About", "/bin/aboutapp", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(58, 104, 190)},
     {"Notepad", "/bin/notepad", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(120, 100, 60)},
+    {"Calculator", "/bin/calc", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(47, 111, 94)},
     {"Doom", "/disk/bin/doomgeneric", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(181, 81, 55)},
 #if DESKTOP_INCLUDE_TEST_APPS
     {"Widgets", "/bin/widgetsdemo", DESKTOP_ICON_DESKTOP, WINDOWD_RGB_LITERAL(96, 110, 140)},